+++ lib/sidekiq/redis_connection.rb
@@ -5,7 +5,7 @@ require 'redis/namespace'
 module Sidekiq
   class RedisConnection
     def self.create(options={})
-      url = options[:url] || ENV['REDISTOGO_URL'] || 'redis://localhost:6379/0'
+      url = options[:url] || determine_redis_provider || 'redis://localhost:6379/0'
       driver = options[:driver] || 'ruby'
       # need a connection for Fetcher and Retry
       size = options[:size] || (Sidekiq.server? ? (Sidekiq.options[:concurrency] + 2) : 5)
@@ -24,5 +24,15 @@ module Sidekiq
       end
     end
     private_class_method :build_client
+
+    def self.determine_redis_provider
+      provider = if ENV.has_key? 'REDISTOGO_URL'
+        'REDISTOGO_URL'
+      else
+        ENV['REDIS_PROVIDER'] || 'REDIS_URL'
+      end
+      ENV[provider]
+    end
+    private_class_method :determine_redis_provider
   end
 end